Fix up source issues that made Dec C unhappy
git-svn-id: https://svn.parrot.org/parrot/trunk@501 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ src/packfile.c
@@ -1711,9 +1711,9 @@ PackFile_Constant_dump(struct PackFile_Constant * self) {
             printf("    [ 'PFC_STRING', {\n");
             printf("        FLAGS    => 0x%04x,\n", self->string->flags);
             printf("        ENCODING => %s,\n",
-                    (long) self->string->encoding->name);
+                    self->string->encoding->name);
             printf("        TYPE     => %s,\n",
-                    (long) self->string->type->name);
+                    self->string->type->name);
             printf("        SIZE     => %ld,\n",
                     (long) self->string->bufused);
             /* TODO: Won't do anything reasonable for most encodings */
